backend/requirements.txt
# Django Backend Requirements
Django==5.2.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
python-decouple==3.8
psycopg2-binary==2.9.9  # For PostgreSQL (optional)
gunicorn==21.2.0  # For production deployment
